Update beetle_SN.py
updated script to prompt user for file to be used instead of hardcoding the file name, added the clear func back in, and adding some info to be printed to the screen including the total word count of the file, total number of unique words found, and how many common words from the list were removed. i will be also removing the code which removes any entries of 5 or less occurrences as the script is shifting to compare 2 files (latest data dump to older data dump) and instead output the difference for each word frequency to better show which words have been mentioned more frequently since the last dataset and which may have gone down. after this i still plan it to make a word cloud with the words who have had the highest uptrend in usage and a bar chat to show more detail.