Update beetle_SN.py
upped the results to the top 500, rotated the x-axis labels to 80-degrees to prevent overlay and line them up better to their tick marks, and have both bar chart and word cloud exports functioning as desired. overall the script is where i want it. more common words might need to be added and the bar chart and word cloud colors might be played with (the word cloud is still randomly choosing the color scheme with each run)